Move QuiltDisplayedError to the gui api package (from the plugin api package).
