+++ src/main/java/org/quiltmc/loader/api/plugin/QuiltPluginContext.java
@@ -20,6 +20,7 @@ import java.nio.file.Path;
 import java.util.concurrent.Callable;
 
 import org.jetbrains.annotations.ApiStatus;
+import org.quiltmc.loader.api.gui.QuiltDisplayedError;
 import org.quiltmc.loader.api.gui.QuiltLoaderText;
 import org.quiltmc.loader.api.plugin.gui.PluginGuiTreeNode;
 import org.quiltmc.loader.api.plugin.solver.LoadOption;
