+++ src/main/java/org/quiltmc/loader/impl/gui/QuiltJsonButton.java
@@ -25,7 +25,7 @@ import org.quiltmc.loader.api.LoaderValue;
 import org.quiltmc.loader.api.LoaderValue.LObject;
 import org.quiltmc.loader.api.gui.QuiltLoaderIcon;
 import org.quiltmc.loader.api.gui.QuiltLoaderText;
-import org.quiltmc.loader.api.plugin.QuiltDisplayedError.QuiltPluginButton;
+import org.quiltmc.loader.api.gui.QuiltDisplayedError.QuiltPluginButton;
 import org.quiltmc.loader.impl.gui.QuiltJsonGui.QuiltBasicButtonAction;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
